requirements.txt
faiss-cpu
numpy
h5py
